# docs/openapi.yaml
openapi: 3.1.0
info:
  title: THCPN IoT Research Platform API
  version: 0.15.0
  description: |
    THCPN 科研物联网数据管理平台后端 API。

    当前文档覆盖已实现的阶段功能：健康检查、短信验证码登录/注册、邮箱验证码验证、密码注册/登录、TOTP MFA、refresh token、logout、会话管理、
    JWT 认证、Workspace 列表/创建、Workspace 成员管理、Project、Site、Device、DeviceRelation、DataStream、DataSource、DataStreamBinding、Telemetry Query、Media Query、Dataset 元信息与遥测查询管理和 ExportJob 导出任务、
    AccessGrant 外部分享/售后临时授权、Invitation、AuditLog、Camera 实时预览，以及系统管理员后台 DataSource / THCPN 标准站和组网站拓扑同步接口。

    正式业务接口优先使用 `Authorization: Bearer <token>`。
    开发环境可通过配置启用 `X-User-ID` fallback；生产环境应关闭该能力。
servers:
  - url: http://127.0.0.1:8080
    description: Local development API server
...
paths:
...
  /api/v1/objects/download:
    get:
      tags:
        - Objects
      operationId: downloadSignedObject
      summary: Download a signed object
      description: Serves temporary HMAC-signed object URLs generated by media and export APIs, primarily for local file object storage. OSS deployments normally receive native OSS pre-signed URLs unless `object_store.public_url_prefix` is configured.
      parameters:
        - name: object_key
          in: query
          required: true
          schema:
            type: string
        - name: expires
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: signature
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object bytes.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/sms/send:
    post:
      tags:
        - Auth
      operationId: sendSmsCode
      summary: Send SMS verification code
      description: |
        Sends a verification code to a normalized phone number.
        With `sms.provider=log`, the code is written to the application log instead of an external SMS provider.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendSmsRequest"
            examples:
              default:
                value:
                  phone: "13800000000"
      responses:
        "200":
          description: Verification code accepted for sending.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendSmsResponse"
              examples:
                sent:
                  value:
                    sent: true
                    expires_in: 300
                    cooldown_seconds: 60
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/email/send:
    post:
      tags:
        - Auth
      operationId: sendEmailVerificationCode
      summary: Send email verification code
      description: |
        Sends a verification code to the current authenticated user's email address.
        With `email.provider=log`, the code is written to the application log instead of an external email provider.
      security:
        - bearerAuth: []
        - devUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailVerificationRequest"
            examples:
              default:
                value:
                  email: zhangsan@example.com
      responses:
        "200":
          description: Verification code accepted for sending.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendCodeResponse"
              examples:
                sent:
                  value:
                    sent: true
                    expires_in: 300
                    cooldown_seconds: 60
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/email/verify:
    post:
      tags:
        - Auth
      operationId: verifyEmailCode
      summary: Verify current user's email
      description: Verifies an email code and marks the current authenticated user's matching email address as verified.
      security:
        - bearerAuth: []
        - devUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
            examples:
              default:
                value:
                  email: zhangsan@example.com
                  code: "123456"
      responses:
        "200":
          description: Email verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyEmailResponse"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/mfa:
    get:
      tags:
        - Auth
      operationId: getMfaStatus
      summary: Get MFA status
      description: Returns whether the current authenticated user has TOTP MFA enabled.
      security:
        - bearerAuth: []
        - devUserHeader: []
      responses:
        "200":
          description: MFA status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MfaStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/mfa/totp/setup:
    post:
      tags:
        - Auth
      operationId: setupTotpMfa
      summary: Generate TOTP setup secret
      description: Generates an encrypted pending TOTP MFA secret for the current authenticated user and returns the setup secret and otpauth URI once.
      security:
        - bearerAuth: []
        - devUserHeader: []
      responses:
        "200":
          description: TOTP setup generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpSetupResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/mfa/totp/enable:
    post:
      tags:
        - Auth
      operationId: enableTotpMfa
      summary: Enable TOTP MFA
      description: Enables TOTP MFA after verifying a code from the pending setup secret.
      security:
        - bearerAuth: []
        - devUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaCodeRequest"
            examples:
              default:
                value:
                  code: "123456"
      responses:
        "200":
          description: TOTP MFA enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MfaStatusResponse"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/mfa/totp:
    delete:
      tags:
        - Auth
      operationId: disableTotpMfa
      summary: Disable TOTP MFA
      description: Disables TOTP MFA after verifying a current TOTP code.
      security:
        - bearerAuth: []
        - devUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaCodeRequest"
            examples:
              default:
                value:
                  code: "123456"
      responses:
        "204":
          description: TOTP MFA disabled.
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/sms/login:
    post:
      tags:
        - Auth
      operationId: loginWithSms
      summary: Login or register with SMS verification code
      description: |
        Verifies the SMS code and issues a JWT access token plus a refresh token.
        If the phone number does not belong to an active user, a new user, personal workspace and Owner membership are created.
        Existing users with TOTP MFA enabled must also supply `mfa_code`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsLoginRequest"
            examples:
              default:
                value:
                  phone: "13800000000"
                  code: "123456"
                  mfa_code: "654321"
                  name: 张三
      responses:
        "200":
          description: SMS login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              examples:
                loggedIn:
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: wJalrXUtnFEMI-K7MDENG_bPxRfiCYEXAMPLE
                    token_type: Bearer
                    expires_in: 86400
                    refresh_expires_in: 2592000
                    user:
                      id: 018fa0ec-9b6e-7a3d-9c7b-1d8b04f31988
                      name: 张三
                      phone: "13800000000"
                      status: active
                      phone_verified_at: "2026-06-30T14:12:54+08:00"
                      last_login_at: "2026-06-30T14:12:54+08:00"
                    created: true
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/password/register:
    post:
      tags:
        - Auth
      operationId: registerWithPassword
      summary: Register with password
      description: |
        Creates a user, personal workspace, Owner membership and password credential in one transaction,
        then issues a JWT access token plus a refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordRegisterRequest"
            examples:
              default:
                value:
                  name: 张三
                  phone: "13800000000"
                  email: zhangsan@example.com
                  password: Str0ngPassword!
      responses:
        "201":
          description: Password registration succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              examples:
                registered:
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: wJalrXUtnFEMI-K7MDENG_bPxRfiCYEXAMPLE
                    token_type: Bearer
                    expires_in: 86400
                    refresh_expires_in: 2592000
                    user:
                      id: 018fa0ec-9b6e-7a3d-9c7b-1d8b04f31988
                      name: 张三
                      phone: "13800000000"
                      email: zhangsan@example.com
                      status: active
                      last_login_at: "2026-06-30T14:12:44+08:00"
                    created: true
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
...
  /api/v1/auth/sessions:
    get:
      tags:
        - Auth
      operationId: listAuthSessions
      summary: List current auth sessions
      description: Lists active refresh sessions for the current authenticated user.
      security:
        - bearerAuth: []
        - devUserHeader: []
      responses:
        "200":
          description: Active auth sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token returned by SMS or password login APIs.
    devUserHeader:
      type: apiKey
      in: header
      name: X-User-ID
      description: Development-only fallback authentication. Production should disable it.
...
  responses:
    InvalidArgument:
      description: Invalid request argument or body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalid:
              value:
                error:
                  code: INVALID_ARGUMENT
                  message: invalid request body
                  request_id: req_123
...
  schemas:
...
    SendSmsRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          $ref: "#/components/schemas/Phone"
...
    SendEmailVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          $ref: "#/components/schemas/Email"
...